+++ CHANGELOG.md
@@ -39,6 +39,13 @@ aws lambda invoke --function-name $PREFIX-ReconciliationReportMigration $OUTFILE
     Elasticsearch
   - Update `@cumlus/api/ecs/async-operation` to not update Elasticsearch index when
     reporting status of async operation
+- **CUMULUS-3806**
+  - Update `@cumulus/db/lib/granule.getGranulesByApiPropertiesQuery` to
+    be parameterized and include a modifier on `temporalBoundByCreatedAt`
+  - Remove endpoint call to and all tests for Internal Reconciliation Reports
+    and updated API to throw an error if report is requested
+  - Update Orca reconciliation reports to pull granules for comparison from
+    postgres via `getGranulesByApiPropertiesQuery`
 - **CUMULUS-3833**
   - Added `ReconciliationReportMigration` lambda to migrate ReconciliationReports from DynamoDB
     to Postgres
@@ -46,7 +53,7 @@ aws lambda invoke --function-name $PREFIX-ReconciliationReportMigration $OUTFILE
   - Added `reconciliation_reports` table in RDS, including indexes
   - Created pg model, types, and translation for `reconciliationReports` in `@cumulus/db`
   - Created api types for `reconciliation_reports` in `@cumulus/types/api`
-  - Updated reconciliation reports lambda to write to new RDS table instead of Dynamo 
+  - Updated reconciliation reports lambda to write to new RDS table instead of Dynamo
   - Updated `@cumulus/api/endpoints/reconciliation-reports` `getReport` and `deleteReport` to work with the new RDS table instead of Dynamo
 
 ## [Unreleased]